requirements.txt
motor
apscheduler
emoji
psutil
pymongo
kurigram
requests
tgcrypto
asyncio
python-dotenv
